requirements.txt
aiofiles==0.7.0
blinker==1.4
click==8.0.3
Deprecated==1.2.13
Flask==2.0.2
h11==0.12.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.11.2
hyperframe==6.0.1
itsdangerous==2.0.1
priority==2.0.0
psycopg2==2.9.1
PyGithub==1.55
PyJWT==2.3.0
PyNaCl==1.4.0
pytz==2021.3
Werkzeug==2.0.2
wsproto==1.0.0
